Refactored file structure
