+++ signtogloss/__init__.py
@@ -1,0 +1,2 @@
+import sys
+sys.path.insert(0, '.')
